Extract pop timer tests to their own file
